requirements.txt
python-jobspy>=1.1.0
httpx[http2]
beautifulsoup4
lxml
playwright
pyhunter
python-dotenv
pandas
